making simple dags; adding slack task; updating vars
