+++ tests/combine_index_dag_test.py
@@ -3,11 +3,11 @@ import os
 import unittest
 import airflow
 try:
-    from dev_combine_index_dag import FCDAG as FCDAGDEV
-    from prod_combine_index_dag import FCDAG as FCDAGPROD
+    from dev_combine_index_dag import DAG as FCDAGDEV
+    from prod_combine_index_dag import DAG as FCDAGPROD
 except:
-    from funcake_dags.dev_combine_index_dag import FCDAG as FCDAGDEV
-    from funcake_dags.prod_combine_index_dag import FCDAG as FCDAGPROD
+    from funcake_dags.dev_combine_index_dag import DAG as FCDAGDEV
+    from funcake_dags.prod_combine_index_dag import DAG as FCDAGPROD
 
 class TestCombineIndexDevDAG(unittest.TestCase):
     """Primary Class for Testing the Combine to FunCake Solr Index DAG."""
@@ -27,7 +27,7 @@ class TestCombineIndexDevDAG(unittest.TestCase):
             "create_collection",
             "combine_index",
             "solr_alias_swap",
-            "slack_post_succ"
+            "success_slack_trigger"
             ])
 
     def test_dag_task_order(self):
@@ -36,7 +36,7 @@ class TestCombineIndexDevDAG(unittest.TestCase):
             "create_collection": ["harvest_oai"],
             "combine_index": ["create_collection"],
             "solr_alias_swap": ["combine_index"],
-            "slack_post_succ": ["solr_alias_swap"]
+            "success_slack_trigger": ["solr_alias_swap"]
         }
 
         for task, upstream_tasks in expected_task_deps.items():
@@ -77,7 +77,7 @@ class TestCombineIndexProdDAG(unittest.TestCase):
             "create_collection",
             "combine_index",
             "solr_alias_swap",
-            "slack_post_succ"
+            "success_slack_trigger"
             ])
 
     def test_dag_task_order(self):
@@ -86,7 +86,7 @@ class TestCombineIndexProdDAG(unittest.TestCase):
             "create_collection": ["harvest_oai"],
             "combine_index": ["create_collection"],
             "solr_alias_swap": ["combine_index"],
-            "slack_post_succ": ["solr_alias_swap"]
+            "success_slack_trigger": ["solr_alias_swap"]
         }
 
         for task, upstream_tasks in expected_task_deps.items():
