+++ variables.json
@@ -23,7 +23,7 @@
   "FREE_LIBRARY_XSL_CONFIG": {
     "schematron_filter": "validations/funcake_reqd_fields.sch",
     "schematron_report": "validations/padigital_missing_thumbnailURL.sch",
-    "xsl_branch": "add-dummy-csv-xsl",
+    "xsl_branch": "master",
     "xsl_filename": "transforms/free_library_csv.xsl",
     "xsl_repository": "tulibraries/aggregator_mdx"
   },
